enhancement(4534): made changes to resolve linting issues

+++ libbeat/publisher/queue/diskqueue/reader_loop.go
@@ -126,7 +126,7 @@ func (rl *readerLoop) processRequest(request readerLoopRequest) readerLoopRespon
 		return readerLoopResponse{err: err}
 	}
 
-	targetLength := uint64(request.endPosition - request.startPosition)
+	targetLength := request.endPosition - request.startPosition
 	for {
 		remainingLength := targetLength - byteCount
